move issue tests to another module
